% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airnow_api_getDataSubset.R
\name{airnow_api_getDataSubset}
\alias{airnow_api_getDataSubset}
\title{Download and parse hourly data from the AirNow data API}
\usage{
airnow_api_getDataSubset(
  starttime = NULL,
  endtime = NULL,
  timezone = "UTC",
  pollutant = c("PM2.5"),
  monitorType = c("permanent", "mobile", "both"),
  baseUrl = "https://www.airnowapi.org/aq/data/"
)
}
\arguments{
\item{starttime}{Desired start datetime (ISO 8601).}

\item{endtime}{Desired end datetime (ISO 8601).}

\item{timezone}{Olson timezone used to interpret dates (required).}

\item{pollutant}{One or more EPA AQS criteria pollutants.}

\item{monitorType}{Subset of all monitors to select.}

\item{baseUrl}{Base URL for archived hourly data.}
}
\value{
Tibble of AirNow hourly data.
}
\description{
This function uses the AirNow data webservice to retrieve
subsets of data that do not exceed a maximum data size which causes
errors.

Datetimes can be anything that is understood by
\code{MazamaCoreUtils::parseDatetime()}. For non-POSIXct values, the
recommended format is "YYYY-mm-dd HH:00:00" or just "YYYYmmddhh".
}
\examples{
\dontrun{
library(AirMonitorIngest)

tbl <-
  airnow_api_getData(
    starttime = 2021101200,
    endtime = 2021101200,
    timezone = "America/Los_Angeles",
    pollutant = "PM2.5",
    monitorType = "permanent"
   )

}
}
