% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airnow_createMeta.R
\name{airnow_createMeta}
\alias{airnow_createMeta}
\title{Create a 'meta' dataframe with required monitor metadata}
\usage{
airnow_createMeta(
  locationTbl = NULL,
  distanceThreshold = NULL,
  airnow_data = NULL
)
}
\arguments{
\item{locationTbl}{Table of "known locations" produced with \pkg{MazamaLocationUtils}.}

\item{distanceThreshold}{Separation distance in meters between "known locations".}

\item{airnow_data}{Table of monitor data obtained with \code{epa_api_getData()}.}
}
\value{
Tibble of device-deployment metadata.
}
\description{
Create a \code{meta} dataframe with AirNow monitor metadata appropriate
for use with the \pkg{MazamaTimeSeries} package.

The data model is that monitor metadata are stored in a tibble named \code{meta}.
with a \code{deviceDeploymentID} unique identifier that is matched by column
names in an associated \code{data} file.
}
