% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airnow_getDataCustom.R
\name{airnow_getDataCustom}
\alias{airnow_getDataCustom}
\title{Download and parse custom data subset from the AirNow data API}
\usage{
airnow_getDataCustom(
  parameterName = c("PM2.5"),
  starttime = NULL,
  endtime = NULL,
  timezone = "UTC",
  bbox = NULL,
  monitorType = c("both", "permanent", "mobile"),
  instrumentType = c("both", "monitor", "sensor"),
  baseUrl = "https://www.airnowapi.org/aq/data/"
)
}
\arguments{
\item{parameterName}{One or more EPA AQS criteria parameter names.}

\item{starttime}{Desired start datetime (ISO 8601).}

\item{endtime}{Desired end datetime (ISO 8601).}

\item{timezone}{Olson timezone used to interpret dates (required).}

\item{bbox}{Bounding box vector defined as c(w, s, e, n).}

\item{monitorType}{Specifies mobile or permanent monitor deployments.}

\item{instrumentType}{Specifies regulatory monitors or low cost sensors.}

\item{baseUrl}{Base URL for archived hourly data.}
}
\value{
Tibble of AirNow hourly data.
}
\description{
This function uses the AirNow data webservice to retrieve
subsets of data. This function is useful when users want data for a small
subset of monitors within a particular region. It can also be used to access
international data for countries outside the Canada-US-Mexico default region.

Datetimes can be anything that is understood by
\code{MazamaCoreUtils::parseDatetime()}. For non-POSIXct values, the
recommended format is "YYYY-mm-dd HH:00:00" or just "YYYYmmddhh".
}
\examples{
\donttest{
library(AirMonitorIngest)

# Fail gracefully if any resources are not available
try({

  setAPIKey("airnow", "<YOUR_API_KEY>")

  # Guatemala City
  monitor <-
    airnow_getDataCustom(
      parameterName = "PM2.5",
      starttime = 20210101,
      endtime = 20220101,
      timezone = "UTC",
      bbox = c(-91.4, 13.7, -89.4, 15.7)
    )

}, silent = FALSE)
}

}
