% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airnow_getHourlyData.R
\name{airnow_getHourlyData}
\alias{airnow_getHourlyData}
\title{Download and parse hourly data from AirNow}
\usage{
airnow_getHourlyData(
  datestamp = NULL,
  baseUrl = "https://files.airnowtech.org/airnow",
  quiet = TRUE
)
}
\arguments{
\item{datestamp}{Integer or character representing YYYYmmddHH.}

\item{baseUrl}{Base URL for archived hourly data.}

\item{quiet}{Logical passed on to
\code{readr::read_delim(progress = !quiet)}.}
}
\value{
Tibble of AirNow hourly data.
}
\description{
The \url{https://airnowtech.org} site provides both air
pollution monitoring data as well as monitoring site location metadata. This
function retrieves and parses a single, hourly data file and returns it as a
tibble.
}
\note{
As of 2016-12-27, it appears that hourly data are available only for
2016 and not for earlier years.

Data from locations whose timezones have a fractional offset from UTC
are removed as the AirMonitor data model only supports data reported on hour
boundaries. As of 2019-06-26, this only applies to US Department of State
monitors in Myanmar, Sri Lanka, India and Nepal.
}
\examples{
\dontrun{
tbl <- airnow_downloadHourlyData(2021101200)
}
}
