% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airnow_getSites.R
\name{airnow_getSites}
\alias{airnow_getSites}
\title{Download and parse AirNow sites metadata}
\usage{
airnow_getSites(
  baseUrl = "https://files.airnowtech.org/airnow/today/",
  sitesFile = "monitoring_site_locations.dat",
  quiet = TRUE
)
}
\arguments{
\item{baseUrl}{URL of the AirNow monitoring site locations file.}

\item{sitesFile}{Name of the AirNow monitoring site locations file.}

\item{quiet}{Logical passed on to
\code{readr::read_delim(progress = !quiet)}.}
}
\value{
Tibble of AirNow site metadata.
}
\description{
The \url{https://airnowtech.org} site provides both air pollution
monitoring data as well as monitoring site location metadata. This function
retrieves the most recent version of the site location metadata file and
returns it as a tibble.

A description of the data format is publicly available at the
\href{https://www.airnowapi.org/docs/MonitoringSiteFactSheet.pdf}{Monitoring Site Fact Sheet}.
}
\note{
As of October, 2021, the description above is incorrect. See the source
code for this function for a correct set of field names.

As of October, 2021, the \code{monitoring_site_locations.dat} file has
an encoding of "CP437" (aka "Non-ISO extended-ASCII" or "IBMPC 437") and will
be converted to "UTF-8" so that French and Spanish language place names are
properly encoded in the returned dataframe.
}
\examples{
\dontrun{
library(AirMonitorIngest)

# Create a directory specifically for AirNow data
dir.create("~/Data/AirNow", recursive = TRUE)

# Set logging level so messages and errors will appear in the console
MazamaCoreUtils::initializeLogging(logDir = "~/Data/AirNow")
logger.setLevel(TRACE)

# Download and parse site metadata
airnow_sites <- airnow_getSites()
}
}
