% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airnow_openWebPages.R
\name{airnow_openWebPages}
\alias{airnow_openWebPages}
\title{Browse AirNow web pages}
\usage{
airnow_openWebPages()
}
\description{
Open the relevant AirNow web pages in the default web browser. These pages
need to be regularly consulted when working with AirNow data and it is helpful
to open them quickly.

\itemize{
\item{\url{https://www.airnowapi.org/docs/MonitoringSiteFactSheet.pdf}}
\item{\url{https://docs.airnowapi.org/docs}}
\item{\url{https://docs.airnowapi.org/files}}
}
}
