% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airnow_update.R
\name{airnow_update}
\alias{airnow_update}
\title{Update AirNow data}
\usage{
airnow_update(
  collectionName = NULL,
  distanceThreshold = 100,
  parameterName = c("PM2.5")
)
}
\arguments{
\item{collectionName}{Character identifier for this table.}

\item{distanceThreshold}{Separation distance in meters between "known locations".}

\item{parameterName}{An EPA AQS criteria parameter name.}
}
\description{
Download AirNow sites metadata and monitoring data, update the known locations
table and return a list with the "meta" and "data" tables filled with recent
data.
}
