% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airnow_updateKnownLocations.R
\name{airnow_updateKnownLocations}
\alias{airnow_updateKnownLocations}
\title{Update the AirNow "known locations" table}
\usage{
airnow_updateKnownLocations(
  locationTbl = NULL,
  distanceThreshold = 100,
  parameterName = c("PM2.5")
)
}
\arguments{
\item{locationTbl}{Table of "known locations".}

\item{distanceThreshold}{Separation distance in meters between "known locations".}

\item{parameterName}{An EPA AQS criteria parameter name.}
}
\value{
Tibble of unique "known locations".
}
\description{
Update a "known locations" table with AirNow monitor metadata.

The latest "sites metadata" file is downloaded from AirNow and cleaned up so
that records can be added to the "known locations" table identified by
\code{collectionName}.

This file is then available for use during the creation of 'meta' dataframes
used in \emph{monitor} objects in the \pkg{MazamaTimeSeries} package.
}
