% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airsis_downloadData.R
\name{airsis_downloadData}
\alias{airsis_downloadData}
\title{Download AIRSIS data}
\usage{
airsis_downloadData(
  startdate = strftime(lubridate::now(tzone = "UTC"), "\%Y0101", tz = "UTC"),
  enddate = strftime(lubridate::now(tzone = "UTC"), "\%Y\%m\%d", tz = "UTC"),
  timezone = "UTC",
  provider = NULL,
  unitID = NULL,
  baseUrl = "http://xxx.airsis.com/vision/common/CSVExport.aspx"
)
}
\arguments{
\item{startdate}{Desired start date (ISO 8601).}

\item{enddate}{Desired end date (ISO 8601).}

\item{timezone}{Olson timezone used to interpret dates (required).}

\item{provider}{Identifier used to modify baseURL \code{[APCD|USFS|...]}}

\item{unitID}{AIRSIS station identifier (will be upcased).}

\item{baseUrl}{Base URL for data queries.}
}
\value{
String containing AIRSIS output.
}
\description{
Download data for a particular station for the desired time period.
Data are returned as a single character string containing the AIRIS output.
}
