% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airsis_getCurrentUnitIDs.R
\name{airsis_getCurrentUnitIDs}
\alias{airsis_getCurrentUnitIDs}
\title{Get AIRSIS unit IDs from the daily summary page}
\usage{
airsis_getCurrentUnitIDs(url = "~/Downloads/apcd_stattable.aspx.html")
}
\arguments{
\item{url}{URL with daily summary links. (NOTE: only file paths are supported.)}
}
\value{
Vector of AIRSIS unit IDs.
}
\description{
Parse the AIRSIS Current Status page for a particular provider
to extract unit IDs for those monitors that are currently reporting.
}
