% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airsis_parseData.R
\name{airsis_parseData}
\alias{airsis_parseData}
\title{Parse AIRSIS data string}
\usage{
airsis_parseData(fileString = NULL, codeDir = NULL)
}
\arguments{
\item{fileString}{Character string containing AIRSIS data.}

\item{codeDir}{Directory where AIRSIS data parsing scripts are found.}
}
\value{
Dataframe of AIRSIS raw monitor data.
}
\description{
Raw character data from AIRSIS are parsed into a tibble.
The incoming \code{fileString} can be read in directly from AIRSIS using
\code{airsis_downloadData()} or from a local file using
\code{readr::read_file()}.
}
\details{
AIRSIS provides monitor data in an extremely raw fashion and does
minimal harmonizing of data. Every year, several new, slightly different data
formats become available. For this reason, all AIRSIS parsing code is
kept outside this package and must exist in \code{codeDir}. This allows for
quick addition of support for new formats without having to rebuild the
package and any systems dependent upon this package.
}
