% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterByDistance.R
\name{clusterByDistance}
\alias{clusterByDistance}
\title{Add distance-clustering information to a dataframe}
\usage{
clusterByDistance(
  tbl,
  clusterDiameter = 1000,
  lonVar = "longitude",
  latVar = "latitude",
  maxClusters = 50
)
}
\arguments{
\item{tbl}{Tibble with geolocation information.}

\item{clusterDiameter}{Diameter in meters used to determine the number of
clusters (see description).}

\item{lonVar}{Name of longitude variable in the incoming tibble.}

\item{latVar}{Name of the latitude variable in the incoming tibble.}

\item{maxClusters}{Maximum number of clusters to try.}
}
\value{
Input tibble with additional columns: \code{clusterLon, clusterLat}.
}
\description{
Distance clustering is used to identify unique deployments.

Standard \code{kmeans} clustering does not work well when clusters can have widely
differing numbers of members. A much better result is acheived with
the Partitioning Around Medoids method available in \code{cluster::pam()}.

The value of \code{clusterDiameter} is compared with the output of
\code{cluster::pam(...)$clusinfo[,'av_diss']} to determine the number of clusters.
}
\references{
\href{https://working-with-data.mazamascience.com/2021/07/15/when-k-means-clustering-fails/}{When k-means clustering fails}
}
