% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_aqs_createData.R
\name{epa_aqs_createData}
\alias{epa_aqs_createData}
\title{Create a 'data' dataframe with required data}
\usage{
epa_aqs_createData(AQS_data = NULL, meta = NULL, parameterCode = NULL)
}
\arguments{
\item{AQS_data}{Table of AQS monitoring data produced with \pkg{epa_aqs_parseHourlyData()}.}

\item{meta}{Table of monitor metadata.}

\item{parameterCode}{EPA "Parameter Code".}
}
\value{
Restructured tibble  with data organized as timeseries columns.
}
\description{
Create a \code{data} dataframe with EPA AQS monitoring data appropriate
for use with the \pkg{MazamaTimeSeries} package.

The data model is that data are stored in a tibble named \code{data} with an initial
\code{datetime} column and monitor-specific data in columns identified with a
\code{deviceDeploymentID} that matches one found in the incoming \code{meta} table
}
