% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_aqs_createLocalArchive.R
\name{epa_aqs_createLocalArchive}
\alias{epa_aqs_createLocalArchive}
\title{Create a 'data' dataframe with required data}
\usage{
epa_aqs_createLocalArchive(
  sites_locationTbl = NULL,
  parameterCode = NULL,
  years = NULL,
  archiveBaseDir = NULL,
  downloadDir = NULL,
  quiet = FALSE
)
}
\arguments{
\item{sites_locationTbl}{EPA AQS "known locations" table for \code{parameterCode} monitors.}

\item{parameterCode}{EPA "Parameter Code".}

\item{years}{Vector of years for which to transform data.}

\item{archiveBaseDir}{Path under which the directory structure will be created.}

\item{downloadDir}{Directory in which to save downloaded EPA .zip files. If \code{NULL}, a temp directory is used.}

\item{quiet}{Logical passed on to \code{utils::download.file()}.}
}
\value{
Returns 0 on success.
}
\description{
Create a local archive of  EPA AQS transofrmed into data appropriate
for use with the \pkg{MazamaTimeSeries} package.

The data model is that data are stored in a tibble named \code{data} with an initial
\code{datetime} column and monitor-specific data in columns identified with a
\code{deviceDeploymentID} that matches one found in the incoming \code{meta} table.
Both \code{data} and \code{meta} will be saved as `.rda`, `.csv` and `.csv.gz`
files for maximum portability to other languages like python, javascript or
even to desktop spreadsheet software.

The archive structure underneath \code{archiveBaseDir} will be:

\preformatted{
<archiveBaseDir>/EPA_AQS/<parameterCode>/<year>/
}

Files will be named as:

\preformatted{
epa_aqs_<parameterCode>_<year>_meta.[rda|csv|csv.gz]
epa_aqs_<parameterCode>_<year>_data.[rda|csv|csv.gz]
}
}
\note{
Depending on your internet connection speeds, downloading files from the EPA
site can take several minutes per file. If you see warning messages that you have
timed out, please download files manually before running this function. This
function will then discover the downloaded files in the \code{downloadDir}.
}
