% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_aqs_getCodes.R
\name{epa_aqs_getCodes}
\alias{epa_aqs_getCodes}
\title{Download and parse tables of EPA AQS codes}
\usage{
epa_aqs_getCodes(
  tableName = NULL,
  baseUrl = "https://aqs.epa.gov/aqsweb/documents/codetables/",
  quiet = TRUE
)
}
\arguments{
\item{tableName}{Character name associated with the code table.}

\item{baseUrl}{Character base URL for EPA AQS code tables.}

\item{quiet}{Logical passed on to \code{readr::read_csv(progress = !quiet)}.}
}
\value{
Tibble of EPA AQS codes.
}
\description{
Download and parse code tables from the US EPA.

See https://www.epa.gov/aqs/aqs-code-list for a list of tables. the
\code{tableName} can be determined by hovering over each link. For example,
the \code{tableName} for the "General > Parameters" table is "parameters".
}
\note{
Column types are determined automatically by \code{readr::read_csv()}
based on the first 1000 rows of input. In addition, any column matching
\code{" Code$"} will be converted to 'character' type (typically from
'numeric'). It is up to the user to review column types
(with \code{dplyr::glimpse()}) and reassign where desired.
}
\examples{
\dontrun{
library(AirMonitorIngest)

# Create a directory specifically for EPA data
dir.create("~/Data/EPA", recursive = TRUE)

# Set logging level so messages and errors will appear in the console
MazamaCoreUtils::initializeLogging(logDir = "~/Data/EPA/")
logger.setLevel(TRACE)

# Get a table of parameter codes
parameter_codes <- epa_aqs_getCodes(tableName = "parameters", quiet = FALSE)
}
}
\references{
\href{https://www.epa.gov/aqs/aqs-code-list}{AQS Codes and Descriptions}
}
