% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_aqs_getFormatDescriptionTables.R
\name{epa_aqs_getFormatDescriptionTables}
\alias{epa_aqs_getFormatDescriptionTables}
\title{Download EPA air quality format description tables}
\usage{
epa_aqs_getFormatDescriptionTables(
  url = "https://aqs.epa.gov/aqsweb/airdata/FileFormats.html"
)
}
\arguments{
\item{url}{Character URL for EPA AQS file format descriptions.}
}
\value{
Named list of tables with format descriptions.
}
\description{
Download tables with descriptions of data file formats. These are needed to
properly interpret the data files downloaded from EPA. Tables are returned as
a named list and include (on 2021-09-24):

\itemize{
\item{"Site Description"}
\item{"Monitor Description"}
\item{"Annual Summary"}
\item{"Daily Summary"}
\item{"Hourly Data"}
\item{"8-Hour Average"}
\item{"Blanks Data"}
}

For example, the `epa_aqs_parseHourlyData()` function uses the information in
the table named "Hourly Data".
}
\examples{
\dontrun{
library(AirMonitorIngest)

formatDescriptionTables <- epa_aqs_getFormatDescriptionTables()
head(formatDescriptionTables$`Hourly Data`)
}
}
\references{
\href{https://aqs.epa.gov/aqsweb/airdata/FileFormats.html}{EPA AirData File Formats page}
}
