% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_aqs_getMonitors.R
\name{epa_aqs_getMonitors}
\alias{epa_aqs_getMonitors}
\title{Download and parse EPA AQS monitor metadata}
\usage{
epa_aqs_getMonitors(
  downloadDir = tempdir(),
  baseUrl = "https://aqs.epa.gov/aqsweb/airdata/",
  quiet = TRUE
)
}
\arguments{
\item{downloadDir}{Directory where .zip file will be saved.}

\item{baseUrl}{Character base URL for the EPA AQS archive.}

\item{quiet}{Logical passed on to \code{utils::download.file()}.}
}
\value{
Tibble of EPA site metadata.
}
\description{
Download monitor metadata from the US EPA and save it to a directory. Then
uncompress and parse the data into a tibble.
}
\examples{
\dontrun{
library(AirMonitorIngest)

# Create a directory specifically for EPA data
dir.create("~/Data/EPA", recursive = TRUE)

# Set logging level so messages and errors will appear in the console
MazamaCoreUtils::initializeLogging(logDir = "~/Data/EPA/")
logger.setLevel(TRACE)

# Download and parse site metadata
AQS_monitors <- epa_aqs_getMonitors(downloadDir = "~/Data/EPA", quiet = FALSE)
}
}
\references{
\href{https://aqs.epa.gov/aqsweb/airdata/download_files.html#Meta}{Site and Monitor Descriptions}
}
