% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_aqs_openWebPages.R
\name{epa_aqs_openWebPages}
\alias{epa_aqs_openWebPages}
\title{Browse EPA AQS web pages}
\usage{
epa_aqs_openWebPages()
}
\description{
Open the relevant EPA AQS web pages in the default web browser. These pages
need to be regularly consulted when working with AQS data and it is helpful
to open them quickly.

\itemize{
\item{\url{https://www.epa.gov/outdoor-air-quality-data/interactive-map-air-quality-monitors}}
\item{\url{https://www.epa.gov/criteria-air-pollutants/naaqs-table}}
\item{\url{https://www.epa.gov/aqs/aqs-code-list}}
\item{\url{https://aqs.epa.gov/aqsweb/airdata/FileFormats.html}}
\item{\url{https://aqs.epa.gov/aqsweb/airdata/download_files.html}}
}
}
