% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_aqs_parseHourlyData.R
\name{epa_aqs_parseHourlyData}
\alias{epa_aqs_parseHourlyData}
\title{Parse hourly EPA air quality data}
\usage{
epa_aqs_parseHourlyData(zipFile = NULL, quiet = TRUE)
}
\arguments{
\item{zipFile}{Absolute path to monitoring data .zip file}

\item{quiet}{Logical passed on to \code{readr::read_csv(progress = !quiet)}.}
}
\value{
Tibble of EPA air quality data.
}
\description{
Uncompress previously downloaded, hourly air quality .zip files from the
US EPA and parse the data into a tibble.

EPA parameter codes and (start year) include:

\enumerate{
\item{44201}{ -- Ozone (1980)}
\item{42401}{ -- SO2 (1980)}
\item{42101}{ -- CO (1980)}
\item{42602}{ -- NO2 (1980)}
\item{88101}{ -- PM2.5 FRM/FEM (2008)}
\item{88502}{ -- PM2.5 non FRM/FEM (1998)}
\item{81102}{ -- PM10 (1988)}
\item{SPEC}{ -- PM2.5 Speciation(2001)}
\item{PM10SPEC}{ -- PM10 Speciation (1988)}
\item{WIND}{ -- Winds (1980)}
\item{TEMP}{ -- Temperature (1980)}
\item{PRESS}{ -- Barometric Pressure (1980)}
\item{RH_DP}{ -- RH and Dewpoint (1980)}
\item{HAPS}{ -- HAPs (1993)}
\item{VOCS}{ -- VOCs (1980)}
\item{NONOxNOy (1980)}
}
}
\note{
Unzipped CSV files are almost 100X larger than the compressed .zip files.
CSV files are removed after data are read into a tibble.
}
\examples{
\dontrun{
library(AirMonitorIngest)

# Create a directory specifically for EPA data
dir.create("~/Data/EPA", recursive = TRUE)

# Set logging level so messages and errors will appear in the console
MazamaCoreUtils::initializeLogging(logDir = "~/Data/EPA/")
logger.setLevel(TRACE)

# Save the download in ~/Data/EPA
zipFile <- epa_aqs_downloadHourlyData(2008, "88101", "~/Data/EPA/", quiet = FALSE)

# Uncompress and parse into a tibble
tbl <- epa_aqs_parseHourlyData(zipFile)
}
}
\references{
\href{https://aqs.epa.gov/aqsweb/airdata/download_files.html#Raw}{EPA AirData Pre-Generated Data Files}

\href{https://aqs.epa.gov/aqsweb/airdata/FileFormats.html#_hourly_data_files}{file format description}
}
