% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_createMeta.R
\name{wrcc_createMeta}
\alias{wrcc_createMeta}
\title{Create a 'meta' dataframe with required monitor metadata}
\usage{
wrcc_createMeta(
  locationTbl = NULL,
  distanceThreshold = NULL,
  wrcc_data = NULL,
  unitID = NULL
)
}
\arguments{
\item{locationTbl}{Table of "known locations" produced with \pkg{MazamaLocationUtils}.}

\item{distanceThreshold}{Separation distance in meters between "known locations".}

\item{wrcc_data}{Table of monitor data obtained with \code{epa_api_getData()}.}

\item{unitID}{WRCC station identifier (will be upcased).}
}
\value{
Tibble of device-deployment metadata.
}
\description{
Create a \code{meta} dataframe with WRCC monitor metadata appropriate
for use with the \pkg{MazamaTimeSeries} package.

The data model has monitor metadata stored in a tibble named \code{meta}.
with a \code{deviceDeploymentID} unique identifier that is matched by column
names in an associated \code{data} file.
}
