% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_createMetaAndData.R
\name{wrcc_createMetaAndData}
\alias{wrcc_createMetaAndData}
\title{Create 'meta' dnad 'data' dataframes from WRCC data}
\usage{
wrcc_createMetaAndData(
  locationTbl = NULL,
  distanceThreshold = NULL,
  wrcc_data = NULL,
  unitID = NULL
)
}
\arguments{
\item{locationTbl}{Tibble of "known locations" produced with \pkg{MazamaLocationUtils}.}

\item{distanceThreshold}{Separation distance in meters between "known locations".}

\item{wrcc_data}{Tibble of WRCC monitor data after QC and clustering have
been applied.}

\item{unitID}{WRCC station identifier (will be upcased).}
}
\value{
List with two tibbles.
}
\description{
Create a \code{meta} dataframe with WRCC monitor metadata and a \code{data}
dataframe with PM2.5 time series appropriate for use with the
\pkg{MazamaTimeSeries} package.

The data model has monitor metadata stored in a tibble named \code{meta}.
with a \code{deviceDeploymentID} unique identifier that is matched by column
names in an associated \code{data} file.
}
