% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_downloadData.R
\name{wrcc_downloadData}
\alias{wrcc_downloadData}
\title{Download WRCC data}
\usage{
wrcc_downloadData(
  startdate = strftime(lubridate::now(tzone = "UTC"), "\%Y010101", tz = "UTC"),
  enddate = strftime(lubridate::now(tzone = "UTC"), "\%Y\%m\%d23", tz = "UTC"),
  timezone = "UTC",
  unitID = NULL,
  baseUrl = "https://wrcc.dri.edu/cgi-bin/wea_list2.pl"
)
}
\arguments{
\item{startdate}{Desired start date (ISO 8601).}

\item{enddate}{Desired end date (ISO 8601).}

\item{timezone}{Olson timezone used to interpret dates (required).}

\item{unitID}{WRCC station identifier (will be upcased).}

\item{baseUrl}{Base URL for data queries.}
}
\value{
String containing WRCC output.
}
\description{
Download data for a particular station for the desired time period.
Data are returned as a single character string containing the WRCC output.

Monitor unitIDs can be found at https://wrcc.dri.edu/cgi-bin/smoke.pl.
}
\examples{
\dontrun{
# Fail gracefully if any resources are not available
try({

  fileString <- wrcc_downloadData(20150701, 20150930, unitID = 'SM16')

  fileString \%>\%
    readr::read_lines() \%>\%
    head(8)

}, silent = FALSE)
}
}
\references{
\href{https://wrcc.dri.edu/cgi-bin/smoke.pl}{Fire Cache Smoke Monitoring Archive}
}
