% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_getCurrentUnitIDs.R
\name{wrcc_getCurrentUnitIDs}
\alias{wrcc_getCurrentUnitIDs}
\title{Get WRCC unit IDs from the daily summary page}
\usage{
wrcc_getCurrentUnitIDs(
  enddate = lubridate::now(tzone = "UTC"),
  url = "https://wrcc.dri.edu/cgi-bin/smoke.pl"
)
}
\arguments{
\item{enddate}{Desired end date (ISO 8601).}

\item{url}{URL with daily summary links.}
}
\value{
Vector of WRCC unit IDs.
}
\description{
Parse the WRCC daily summary page to extract unit IDs for those
monitors that are currently reporting.
}
\examples{
\dontrun{
# Fail gracefully if any resources are not available
try({

  wrcc_getCurrentUnitIDs()

}, silent = FALSE)
}
}
\references{
\href{https://wrcc.dri.edu/cgi-bin/smoke.pl}{Fire Cache Smoke Monitoring Archive}
}
