% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_getDailyUnitIDs.R
\name{wrcc_getDailyUnitIDs}
\alias{wrcc_getDailyUnitIDs}
\title{Get WRCC unit IDs from the daily summary page}
\usage{
wrcc_getDailyUnitIDs(url = "https://wrcc.dri.edu/cgi-bin/smoke.pl")
}
\arguments{
\item{url}{URL with daily summary links.}
}
\value{
Vector of WRCC unit IDs.
}
\description{
Parse the WRCC daily summary page to extract unit IDs for those
monitors that are currently reporting.
}
\examples{
\dontrun{
# Fail gracefully if any resources are not available
try({

  wrcc_getDailyUnitIDs()

}, silent = FALSE)
}
}
\references{
\href{https://wrcc.dri.edu/cgi-bin/smoke.pl}{Fire Cache Smoke Monitoring Archive}
}
