% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_getUnitIDs.R
\name{wrcc_getUnitIDs}
\alias{wrcc_getUnitIDs}
\title{Get WRCC unit IDs}
\usage{
wrcc_getUnitIDs(
  startdate = lubridate::now(tzone = "UTC"),
  enddate = NULL,
  by = c("day", "week", "month"),
  url = "https://wrcc.dri.edu/cgi-bin/smoke.pl"
)
}
\arguments{
\item{startdate}{Desired start date (ISO 8601).}

\item{enddate}{Desired end date (ISO 8601).}

\item{by}{Increment by day, week or month.}

\item{url}{WRCC URL with daily summary links.}
}
\value{
Vector of WRCC unit IDs.
}
\description{
Parse the WRCC daily web service response to extract unit IDs
for those monitors that were reporting on a particular date.

With no arguments, this function will return the unitIDs of monitors that
are currently reporting.
}
\examples{
\dontrun{
# Fail gracefully if any resources are not available
try({

  wrcc_getUnitIDs()

}, silent = FALSE)
}
}
\references{
\href{https://wrcc.dri.edu/cgi-bin/smoke.pl}{Fire Cache Smoke Monitoring Archive}
}
