% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_parseData.R
\name{wrcc_parseData}
\alias{wrcc_parseData}
\title{Parse WRCC data string}
\usage{
wrcc_parseData(fileString = NULL)
}
\arguments{
\item{fileString}{Character string containing WRCC data.}
}
\value{
Dataframe of WRCC raw monitor data.
}
\description{
Parse raw character data from WRCC into a tibble.
The incoming \code{fileString}
can be read in directly from WRCC using \code{wrcc_downloadData()} or from a
local file using \code{readr::read_file()}.

The type of monitor represented by \code{fileString} is inferred from the
column names using \code{wrcc_identifyDataFormat()} and appropriate column
types are assigned. The character data are then processed, read into a tibble,
and augmented in the following ways:

\enumerate{
\item{Spaces at the beginning and end of each line are moved.}
\item{All header lines beginning with ':' are removed.}
}
}
\examples{
\dontrun{
# Fail gracefully if any resources are not available
try({

  library(AirMonitorIngest)

  fileString <- wrcc_downloadData(20150701, 20150930, unitID = 'SM16')
  tbl <- wrcc_parseData(fileString)

}, silent = FALSE)
}
}
\references{
\href{https://wrcc.dri.edu/cgi-bin/smoke.pl}{Fire Cache Smoke Monitoring Archive}
}
