% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrcc_updateKnownLocations.R
\name{wrcc_updateKnownLocations}
\alias{wrcc_updateKnownLocations}
\title{Update the WRCC known locations table}
\usage{
wrcc_updateKnownLocations(
  wrcc_locationTbl = NULL,
  airnow_locationTbl = NULL,
  distanceThreshold = NULL,
  wrcc_data = NULL
)
}
\arguments{
\item{wrcc_locationTbl}{WRCC table of "known locations".}

\item{airnow_locationTbl}{AirNow table of "known locations".}

\item{distanceThreshold}{Separation distance in meters between "known locations".}

\item{wrcc_data}{Tibble of parsed/QC'ed/clustered WRCC data.}
}
\value{
Tibble of unique "known locations".
}
\description{
Update a "known locations" table with WRCC monitor metadata.

Clustered locations in incoming WRCC data are used to find "known locations"
in existing WRCC and AirNow location tables in that order. If no known locations
are found, new ones are created.

The returned tibble will have all records in the incoming \code{wrcc_locationTbl}
as well as any additional records needed for new locations.

The returned known locations tibble is then available for use during the c
reation of 'meta' dataframes used in \emph{monitor} objects in the
\pkg{MazamaTimeSeries} package.
}
