% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airnow_createData.R
\name{airnow_createData}
\alias{airnow_createData}
\title{Create a 'data' dataframe with monitor data}
\usage{
airnow_createData(
  locationTbl = NULL,
  distanceThreshold = NULL,
  meta = NULL,
  airnow_data = NULL
)
}
\arguments{
\item{locationTbl}{Table of "known locations" produced with \pkg{MazamaLocationUtils}.}

\item{distanceThreshold}{Separation distance in meters between "known locations".}

\item{meta}{Table of metadata produced with \code{airnow_createMeta}.}

\item{airnow_data}{Table of monitor data obtained with \code{airnow_getData()}.}
}
\value{
Tibble of device-deployment metadata.
}
\description{
Create a \code{data} dataframe with AirNow monitor data appropriate
for use with the \pkg{MazamaTimeSeries} package.

The data model has monitor metadata stored in a tibble named \code{meta}.
with a \code{deviceDeploymentID} unique identifier that is matched by column
names in an associated \code{data} file.
}
\note{
Some records in \code{meta} may not be represented in \code{airnow_data}.
Further work is requried to guarantee an exaxt match between \code{meta} and
the tibbles returned by this function.
}
