% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airnow_getDataSubset.R
\name{airnow_getDataSubset}
\alias{airnow_getDataSubset}
\title{Download and parse hourly data from the AirNow data API}
\usage{
airnow_getDataSubset(
  parameterName = c("PM2.5"),
  starttime = NULL,
  endtime = NULL,
  timezone = "UTC",
  monitorType = c("both", "permanent", "mobile"),
  baseUrl = "https://www.airnowapi.org/aq/data/"
)
}
\arguments{
\item{parameterName}{One or more EPA AQS criteria parameter names.}

\item{starttime}{Desired start datetime (ISO 8601).}

\item{endtime}{Desired end datetime (ISO 8601).}

\item{timezone}{Olson timezone used to interpret dates (required).}

\item{monitorType}{Subset of all monitors to select.}

\item{baseUrl}{Base URL for archived hourly data.}
}
\value{
Tibble of AirNow hourly data.
}
\description{
This function uses the AirNow data webservice to retrieve
subsets of data that do not exceed a maximum data size which causes
errors.

Datetimes can be anything that is understood by
\code{MazamaCoreUtils::parseDatetime()}. For non-POSIXct values, the
recommended format is "YYYY-mm-dd HH:00:00" or just "YYYYmmddhh".
}
