% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airnow_getHourlyData.R
\name{airnow_getHourlyData}
\alias{airnow_getHourlyData}
\title{Download and aggregate multiple hourly data files from AirNow}
\usage{
airnow_getHourlyData(
  parameterName = c("PM2.5"),
  starttime = NULL,
  endtime = NULL,
  timezone = "UTC",
  baseUrl = "https://files.airnowtech.org/airnow",
  verbose = FALSE
)
}
\arguments{
\item{parameterName}{One or more EPA AQS criteria parameter names.}

\item{starttime}{Desired start datetime (ISO 8601).}

\item{endtime}{Desired end datetime (ISO 8601).}

\item{timezone}{Olson timezone used to interpret dates (required).}

\item{baseUrl}{Base URL for archived hourly data.}

\item{verbose}{Logical flag controlling detailed progress statements.}
}
\value{
Tibble of AirNow hourly data.
}
\description{
This function downloads, parses and combines hourly data files
from \code{baseUrl}.
}
\note{
Data returned by this function does not include data from "temporary"
monitors and does not include a field for \code{fullAQSID}. For these reasons,
this function is deprecated in favor of \code{\link{airnow_getData}}.

Parameters included in AirNow data include at least the following list:

\enumerate{
\item{BARPR}
\item{BC}
\item{CO}
\item{NO}
\item{NO2}
\item{NO2Y}
\item{NO2X}
\item{NOX}
\item{NOOY}
\item{OC}
\item{OZONE}
\item{PM10}
\item{PM2.5}
\item{PRECIP}
\item{RHUM}
\item{SO2}
\item{SRAD}
\item{TEMP}
\item{UV-AETH}
\item{WD}
\item{WS}
}

Passing a vector of one ore more of the above names as the \code{parameters}
argument will cause the resulting tibble to be filtered to contain only
records for those parameters.

Datetimes can be anything that is understood by
\code{MazamaCoreUtils::parseDatetime()}. For non-POSIXct values, the
recommended format is "YYYY-mm-dd HH:00:00" or just "YYYYmmddhh".

Hourly data are available from AirNow starting on 2014-01-01.

Data from locations whose timezones have a fractional offset from UTC
are removed as the \emph{mts_monitor} data model only supports data reported
on hour boundaries. As of 2019-06-26, fractional offsets only applies to US
Department of State monitors in Myanmar, Sri Lanka, India and Nepal.
}
\examples{
\dontrun{
# Fail gracefully if any resources are not available
try({

library(AirMonitorIngest)

tbl <-
  airnow_getHourlyData(
    parameterName = "PM2.5",
    starttime = 2021101200,
    endtime = 2021101204,
    timezone = "America/Los_Angeles",
    verbose = TRUE
   )

}, silent = FALSE)
}
}
