% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa_aqs_createMTSMeta.R
\name{epa_aqs_createMeta}
\alias{epa_aqs_createMeta}
\title{Create a meta dataframe with required data}
\usage{
epa_aqs_createMeta(locationTbl)
}
\arguments{
\item{locationTbl}{Table of "known locations" produced with \pkg{MazamaLocationUtils}.}
}
\value{
Tibble enhanced with additional columns.
}
\description{
Create a \code{meta} dataframe with monitor-specific metadata appropriate
for use with the \pkg{MazamaTimeSeries} package.

\preformatted{
Core meta columns include:

  deviceDeploymentID -- unique identifier (see MazmaLocationUtils)
  deviceID -- device identifier
  locationID -- location identifier (see MazmaLocationUtils)
  siteName -- English language name
  longitude -- decimal degrees E
  latitude -- decimal degrees N
  elevation -- elevation of station in m
  countryCode -- ISO 3166-1 alpha-2
  stateCode -- ISO 3166-2 alpha-2
  timezone -- Olson time zone

Core data columns include:

  datetime -- measurement time (UTC)
}
}
